refactor: decouple ontology from core, invert dependency
Make nvisy-ontology fully independent (zero internal deps):
- Move ContentSource from core to ontology::entity
- Remove document_types field from RuleCondition
- Remove nvisy-core dependency from ontology

Make nvisy-core depend on nvisy-ontology:
- Re-export ContentSource from core::content for backwards compat
- Add From<ontology::Error> for core::Error bridge
- Stop publishing nvisy-core to crates.io

Update release CI to only publish nvisy-ontology.
