+++ crates/nvisy-ontology/Cargo.toml
@@ -23,9 +23,6 @@ all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
 
 [dependencies]
-# Internal crates
-nvisy-core = { workspace = true, features = [] }
-
 # (De)serialization
 serde = { workspace = true, features = [] }
 serde_json = { workspace = true, features = [] }
