+++ crates/nvisy-ontology/src/context/biometric/face.rs
@@ -1,9 +1,9 @@
 //! Face biometric reference data.
 
-use nvisy_core::content::ContentSource;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
+use crate::entity::ContentSource;
 use crate::math::BoundingBox;
 
 /// Reference face data for identity matching.
