+++ crates/nvisy-ontology/src/context/biometric/voice.rs
@@ -1,9 +1,9 @@
 //! Voice biometric reference data.
 
-use nvisy_core::content::ContentSource;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
+use crate::entity::ContentSource;
 use crate::math::TimeSpan;
 
 /// Reference voice data for speaker identification.
