+++ crates/nvisy-ontology/src/context/document/signature.rs
@@ -1,9 +1,9 @@
 //! Signature reference data.
 
-use nvisy_core::content::ContentSource;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
+use crate::entity::ContentSource;
 use crate::math::BoundingBox;
 
 /// Reference handwritten signature for verification.
