+++ crates/nvisy-ontology/src/context/document/template.rs
@@ -1,9 +1,9 @@
 //! Document template reference data.
 
-use nvisy_core::content::ContentSource;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
+use crate::entity::ContentSource;
 use crate::math::BoundingBox;
 
 /// Reference document template for layout/type classification.
