+++ crates/nvisy-ontology/src/lib.rs
@@ -8,6 +8,7 @@ pub mod math;
 
 mod error;
 pub use self::error::{Error, Result};
+
 pub mod policy;
 pub mod provenance;
 pub mod workflow;
