+++ crates/nvisy-ontology/src/provenance/record/redaction.rs
@@ -1,11 +1,11 @@
 //! Audit-facing redaction record.
 
-use nvisy_core::content::ContentSource;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use super::review::ReviewDecision;
+use crate::entity::ContentSource;
 
 /// An audit-facing record of a redaction that was (or will be) applied.
 ///
